[ refactor ] updateModal & coachMark custom hook을 공통 hook으로 생성
