+++ src/@components/CardCollectionPage/index.tsx
@@ -7,16 +7,16 @@ import Header from "../@common/Header";
 import HeaderMinVer from "../@common/Header/HeaderMinVer";
 import useGTMPage from "../@common/hooks/useGTMPage";
 import useModal from "../@common/hooks/useModal";
+import useModalPopup from "../@common/hooks/useModalPopup";
 import useScroll from "../@common/hooks/useScrollToTop";
-import LoginModal from "../@common/LoginModal";
 import SuspenseBoundary from "../@common/SuspenseBoundary";
 import CardSlider from "./CardSlider";
 import CoachMark from "./CoachMark";
 import FilterModal from "./FilterModal";
 import { useCardLists } from "./hooks/useCardLists";
-import useCoachMark from "./hooks/useCoachMark";
 import useCTAFilter from "./hooks/useCTAFilter";
 import * as St from "./style";
+const COACH_MARK_LOCAL_KEY = "coach-mark-shown";
 
 export default function CardCollectionPage() {
   return (
@@ -36,7 +36,10 @@ function CardCollectionContent() {
 
   const { isModalOpen: isFilterModalOpen, toggleModal: toggleFilterModal } = useModal();
 
-  const { isOpened: isCoachMarkOpen, handleCloseCoachMark: toggleCoachMark } = useCoachMark();
+  const { isOpened: isCoachMarkOpen, handleCloseModal: toggleCoachMark } = useModalPopup(
+    localStorage,
+    COACH_MARK_LOCAL_KEY,
+  );
 
   const isSliderDown = useRecoilValue(isSliderDownState);
